Add a method to include all packages from one repo/mirror to the filter of another mirror.
This is useful for repositories containing updates of partial
mirrors.

For example, consider a vendor who releases a base
repo, and publishes security updates as needed to another repo.
If you want to make a partial mirror of this structure, you will
probably want to create a mirror of $RELEASE based on a filter
with dependencies included.  So you might create a mirror like

aptly mirror create -filter='nginx' -filter-with-deps $BASE_NAME $URL $RELEASE main

Then aptly will pull in nginx and all the packages it depends on.
Now imagine a security update is released to a package nginx depends
on.  How can you create a mirror of the security repo that includes
everything you fetched from the main repo?  You can't use the same
filter since the security repo might not even have nginx in it if
nginx has not gotten a security update yet.  Without any changes
your options are to mirror the whole security updates repo, or
list every transitive dependency in your repo filter query.

With this new option when creating a mirror you can ask that all packages
from a parent repo/mirror are included in a repo's filter.  So for the
example above, you could set up the security mirror as:

aptly mirror create -filter="None" -filter-with-deps -with-mirror-packages=$BASE_NAME $SECURITY_NAME $SECURITY_URL $RELEASE-security main